Use mousedown instead of click in MenuButton to fix issue.
Previously, if you mousedowned on a MenuButton, it would immediately
open the dropdown. If you moved the mouse over the dropdown (or if the
dropdown opened over the button, which will happen if position: 'cover'
is used) and then mouse-upped over the dropdown, then the MenuButton
would close. MenuButton doesn't want to close if the element that was
interacted with was part of the MenuButton or its dropdown, but the
problem is that when a click event is dispatched after a user mousedowns
on one element and mouseups on a different element, the click event is
dispatched on the closest common ancestor of the two elements. In the
case where the user mousedowns on the MenuButton and mouseups on the
dropdown, the closest common ancestor will be the page body, the click
event will be dispatched there, and MenuButton will think something
outside of itself and its dropdown was interacted with. We need to not
use the click event for this then.